LPS-122456 Provides buildFragment from frontend-js-web
To simplify the transition, provide temporarily (or permanently) a
buildFragment utility in line with the one used in frontend-js-spa-web
and update usages.

Most usages can surely be replaced safely by createContextualFragment
but we can do that in a follow up pass.
