+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/index.es.js
@@ -75,6 +75,7 @@ export {default as throttle} from './liferay/throttle.es';
 // Util API
 
 export {default as addParams} from './liferay/util/add_params';
+export {default as buildFragment} from './liferay/util/build_fragment';
 export {default as fetch} from './liferay/util/fetch.es';
 export {default as focusFormField} from './liferay/util/focus_form_field';
 export {default as getPortletId} from './liferay/util/get_portlet_id';
