+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/compat/modal/Modal.es.js
@@ -18,6 +18,7 @@ import dom from 'metal-dom';
 import {EventHandler} from 'metal-events';
 import Soy from 'metal-soy';
 
+import buildFragment from '../../util/build_fragment';
 import templates from './Modal.soy';
 
 const KEY_CODE_ESC = 27;
@@ -236,7 +237,7 @@ class Modal extends Component {
 	 */
 
 	_valueOverlayElementFn() {
-		return dom.buildFragment('<div class="modal-backdrop fade show"></div>')
+		return buildFragment('<div class="modal-backdrop fade show"></div>')
 			.firstChild;
 	}
 }
