+++ modules/apps/frontend-js/frontend-js-web/test/liferay/util/form/get_form_element.es.js
@@ -12,15 +12,12 @@
  * details.
  */
 
-'use strict';
-
-import dom from 'metal-dom';
-
+import buildFragment from '../../../../src/main/resources/META-INF/resources/liferay/util/build_fragment';
 import getFormElement from '../../../../src/main/resources/META-INF/resources/liferay/util/form/get_form_element.es';
 
 describe('Liferay.Util.getFormElement', () => {
 	it('returns null if the form parameter is not a form node', () => {
-		const fragment = dom.buildFragment('<div />');
+		const fragment = buildFragment('<div />');
 
 		const form = fragment.firstElementChild;
 
@@ -29,7 +26,7 @@ describe('Liferay.Util.getFormElement', () => {
 	});
 
 	it('returns null if the elementName parameter is not a string', () => {
-		const fragment = dom.buildFragment('<form />');
+		const fragment = buildFragment('<form />');
 
 		const form = fragment.firstElementChild;
 
@@ -38,7 +35,7 @@ describe('Liferay.Util.getFormElement', () => {
 	});
 
 	it('returns null if the element does not exist withing the form', () => {
-		const fragment = dom.buildFragment(`
+		const fragment = buildFragment(`
 					<form data-fm-namespace="_com_liferay_test_portlet_" id="fm">
 						<input name="_com_liferay_test_portlet_foo" type="text" value="abc">
 					</form>
@@ -50,7 +47,7 @@ describe('Liferay.Util.getFormElement', () => {
 	});
 
 	it('returns element value if the element does exist withing the form', () => {
-		const fragment = dom.buildFragment(`
+		const fragment = buildFragment(`
 					<form data-fm-namespace="_com_liferay_test_portlet_" id="fm">
 						<input name="_com_liferay_test_portlet_foo" type="text" value="abc">
 					</form>
