+++ modules/apps/map/map-common/package.json
@@ -1,7 +1,7 @@
 {
 	"dependencies": {
+		"frontend-js-web": "*",
 		"metal": "2.16.8",
-		"metal-dom": "2.16.8",
 		"metal-state": "2.16.8"
 	},
 	"name": "map-common",
