+++ modules/apps/map/map-common/src/main/resources/META-INF/resources/js/MapBase.es.js
@@ -12,7 +12,7 @@
  * details.
  */
 
-import {buildFragment} from 'metal-dom';
+import {buildFragment} from 'frontend-js-web';
 import State, {Config} from 'metal-state';
 
 import GeoJSONBase from './GeoJSONBase.es';
